Rename PageflowTooltip to Tooltip
