+++ entry_types/scrolled/package/spec/frontend/Tooltip-spec.js
@@ -1,14 +1,14 @@
 import React from "react"
 import {render as testingLibraryRenderer} from '@testing-library/react'
 
-import {PageflowTooltip} from "frontend/PageflowTooltip";
+import {Tooltip} from "frontend/Tooltip";
 
-describe("PageflowTooltip", () => {
+describe("Tooltip", () => {
   const render = props =>
     testingLibraryRenderer(
-      <PageflowTooltip fixed={props.fixed} content={props.content} closeOther={props.closeOther} classWhenOpen={props.classWhenOpen}>
+      <Tooltip fixed={props.fixed} content={props.content} closeOther={props.closeOther} classWhenOpen={props.classWhenOpen}>
         <button data-testid={props.triggerId || 'tooltip_trigger'} >Tooltip trigger</button>
-      </PageflowTooltip>
+      </Tooltip>
     )
 
   it("should render the content", () => {
