+++ entry_types/scrolled/package/src/frontend/Tooltip/components/Arrow/index.js
