+++ entry_types/scrolled/package/src/frontend/Tooltip/components/Bubble/Bubble.module.css
