+++ entry_types/scrolled/package/src/frontend/Tooltip/components/Bubble/index.js
